requirements.txt
scrapy
psycopg2-binary
scrapy-camoufox
cleantext
